pyproject.toml
[tool.poetry]
name = "f-data-zone-bk"
version = "0.1.0"
description = ""
authors = ["ZalivkoDenis <ZalivkoDenis@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.29.0"
black = "^24.4.2"
pydantic = {extras = ["email"], version = "^2.7.1"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"
asyncio = "^3.4.3"
psycopg = {extras = ["binary"], version = "^3.1.18"}
python-dateutil = "^2.9.0.post0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
bcrypt = "^4.1.3"
python-multipart = "^0.0.9"
fastapi = "^0.111.0"
pycryptodome = "^3.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
